prez-backend/prez-app/requirements.txt
SQLAlchemy
fastapi
pymysql
uvicorn
cryptography
PyJWT
python-decouple